Initial commit, first cut of spec with 3 basic entities.
We currently have 3 entities:

  1. User
  2. Organization
  3. Document

Each of these 3 entities have barebones attributes, but we shouldn't
need much more for an initial prototype. We still need to flesh out most
of the descriptions and perhaps provide some sensible default values,
but we currently have total validation coverage for the model attributes
that have been defined.

A good next step will be to convert some of these attribute definitions
to Stoplight "Traits" so we can deduplicate definitions of things like
`id` format, name validation, and other things.

Currently, we're not including Stoplight OAS extensions in source
control. That may change in the future. Better yet, we should have a
branch with the extensions.