Rip out GET examples, disallow adtnl props.
Here we refine our specification a little bit to disallow additional
properties on everything. We do this by making all of our *ReadOnly
models _strictly_ duplicates of their corresponding Master models. For
instance, "Docuemnt ReadOnly" is generated by copying "Document" and
then modifying only cursorily (e.g. making all properties required).

It then follows that our Master models ("Document", "Resource", etc.)
are now our single Source of Truth for model definitions. We work around
differences in property requirement (we want to let the client specify
only the parts of an entity they wish to update, but enforce returning
all properties from the server) by copying the master model to a
readonly model (e.g. "document readonly") to define the operation
output, which we then delete and re-copy on an update to the master
model.

We also rip out all of the non-reference model examples to simplify and
avoid forgetting to update/cull out-of-date examples as the API evolves.

Last thing here is to update the populate_api.sh script to conform to
the new representation of the Resource entity/model.