Update OAS spec to require less, enhance population script.
This commit brings some small changes to the bulb OAS, relaxing some of
the unnecessary output property requirements (required props on
*ReadOnly models). Now, only fields that will actually be required by
the client application will be required by the schema/specification.

We also make some shweet enhancements to the populate_api.sh script,
namely adding a -t/--test flag that runs GETs on the bare endpoints for
all of our top-level entities, which has the effect of listing all
entries for that entity. This has the benefit of validating that both

  a.) our implementation conforms to our spec vis-a-vis entity output
  b.) our populate_api.sh script is kept up to date

Also added to the populate_api.sh script are a verbosity flag (-v) and
help/usage funcitonality (-h).