Dockerize the shit out of the bulb API (locally).
This commit wraps the API and its virtualenv up in docker and provides
scripts for a local development environment that is as close to Lambda's
environment as possible. To acheive this, I snagged the Dockerfile from
[here](https://github.com/danielwhatmuff/zappa/blob/master/Dockerfile)
and modified it.

To run the python process inside of the same (as much as possible)
virtualenv as what I have on the mac, the entrypoint invokes whatever
command is specified to the contianer (via `docker run`) using a simple
script `venv_exec`. One can still run the API locally in a virtualenv;
this will not interfere with the docker builds or execuiton.

The user can also run arbitary commands in the API's virtualenv context.
If deploying dependencies compiled for a lambda-like environment is
critical, then `zappa update` should be invoked via e.g.

  `$ /path/to/run_api.sh zappa update`

Then container must be rebuilt any time the `requirements.txt` of the
project are updated in order for those changes to propagate. Before
deploying this contaniner, the developer would need to add logic to copy
the bulb API source into the container's filesystem, as obviously it
wouldn't be possible to portably mount the directory containing the
source via `docker run -v` on arbitrary hosts.

Run bare, the container will call `./run.py local` and present it on
port 8080 of the container. By default, `run_api.sh` also links port
8080 on the container to 8080 on the main host.

Running and building the container is further simplified with direnv:

```
$ cat .envrc
layout python

build() {
    set -x
    local tag=${1:-"lambda"}
    docker build --rm -t $tag .
    set +x
}
export_function build

run_api() {
    ./scripts/docker/run_api.sh $@
}
export_function run_api

export FLASK_DEBUG=1
```

See `run_api.sh --help` for more pertient details/parameters.