All models are now pynamo-backed.
In this commit we migrate the remaining models to be backed by pynamo.
We also consolidate the logic for the various CRUD methods into
functions in handlers.py which refer to the generic models as
"entities". This is the terminology we will use to discuss the BulbModel
inheritants.

We also specify the null=True kwarg in some entity attribute
declarations to tell pynamo to allow None-valued instance attributes to
be serialized to dynamodb without raising an error.

All of the entities' Meta member classes inherit from BulbModel so we
can share things like host configuration and base IOps config. However,
we should override the r/w IOps in entities/tables (there is a 1:1
relation between entities and DDB tables) that are likely to incur
heavier load.

TODO
====
- [ ] test this out on a live dynamodb instance!!
- [ ] write some gaddamn unit tests!!
- [x] move as much of each BulbModel inheritor's Meta class to BulbModel
- [ ] make dynamodb host dependent upon DEBUG configuration (DEBUG->local)
- [x] factor out all table update, return logic, etc into common methods,
  and have each entity's CRUD methods call that.