Docker support for DDB local and prod.
This commit modifies our `run_api.sh` script such that it by default
specifies the `--link` option to `docker run` with a default value of
`ddb_local`. This default assumes that the docker host has another
container on it running dyanamodb with port 8000 being exposed and used
by the ddb process. The user can specify no linkage with `--no-link`.

We also modify `models.py` to attempt getting the ddb_local container's
linked TCP address, changing it to a HTTP URI if successful and falling
back to localhost if not. At some point we should move the name of the
DDB_LOCAL_PORT variable over to a config file.

This change essentially allows us to run the API code locally either in
a container or not, in both cases being able to talk to any of a local
DDB container, a DDB Local instance running on the dev host, or DDB's
prod endpoints. All have been tested both in Docker and not. However,
none of this has been tested on a remote Docker host.

Also, forgot to include updates `requirements.txt` in previous DDB
commits, so adding that in here.