Parametrize create CRUD tests with all Models.
In this commit, we move all of our create API path CRUD tests over to an
xunit-style test class organization, and parametrize each test with all
of our models. This is nice, because we get great coverage for free.

We also split our test data into two categories: miminal and full.
"full" data is meant to have examples that contain all possible
properties. Likewise, "minimal" data has the bare minimum required by
the API. We need to use the minimal data for our bad attr-type test
cases because they work by deleting each attr and asserting that we
return a 400. Any non-necessary attrs would thus get deleted, but we'd
return a 201 because that attr wasn't needed to create the entity.

We should consider eliminating this test class and just using
module-level test functions if we end up using fresh_db fixture for
everything...