Implement user authentication with OAuth2 and JWT.
Overview
========
This commit introduces a new API endpoint `/auth` which is responsible
for issuing and verifying authentication tokens. Tokens are passed on
each request by the client under the `Authorization` header, complying
with the bearer token format of OAuth2. An example using curl:

  curl https://dev.bulb.co/ \
        -H 'Authorization: Bearer 123.my_token.xyz'

The tokens themselves are constructed according to the JSON Web Tokens
(JWT) standard as established in RFCs 7519 and 7523. Bulb's JWT's make 3
(standard) claims:

  - `uid`: user's id in standard UUID4 format
  - `iat`: seconds since epoch UTC at which token was issued
  - `exp`: seconds since epoch UTC at which token expires

Currently, we use `pyjwt`'s default signing algorithm (HMAC 256) to sign
the tokens, and a weak, non-production-ready secret (namely, "yourmom"
lulz). Conveniently, the currently authenticated user's ID can be
accessed and assigned to some variable `uid` like so:

```python
import flask
uid = flask.request.user
```

Authentication
==============
We've decided to use connexion's standard (actually, only) built-in
authentication mechanism, which utilizes the OAuth2 authorization
standard as described in RFC 6749.

Although OAuth2 is billed as an "authorization" framework, it is widely
used for both authentication and authorization. For an in-depth overview
of the subtle, but important, differences between the two, see [this S/O
post](http://bit.ly/2tHwWBm). As of this commit, we use OAuth2 for
**authentication only**. Perhaps in the future, we'll get fancy and
actually use its authrization facilities, but for now, we don't perform
any authorization at all. The most likely immediate route to proper
authorization is goint to be implementing simple checks in
`handlers.py`, calling out to DynamoDB and its indices (probably using
the user's affiliated `organization` as a range key) to check if the
user is _authorized_ to access a given entity. This will be rather crude
and clunky, but it's a start. A more in-depth implementation/integration
of OAuth2 in the future should allow for much more granular access
control.

Connexion relies on the user specifying an authorization URL where an
OAuth2-compliant server is running. The user can do so either in the OAS
specification itself (under the `x-tokenInfoUrl` extension) or with the
environment variable TOKENINFO_URL, as we have chosen to do. To
accomodate our deployment constraints and limit or eliminate external
network calls on our per-request authentication (which would be
massively inefficient as we would be unable to cache results in Lambda),
we implement this server within bulb's API itself and sidecar a second
connexion `app` instance in the same python process. To do this, we use
the standard lib's `threading` library to kick off a second thread
running the token validation instance during the main app's
initialization. This token validation instance's only function is to
listen for GET's on the `/auth` endpoint. We could have broken this out
into a distinct OAS specification and connexion app, but I'm lazy and
this is simpler.  However, to avoid exposing this validation endpoint to
would-be assailants, we do two things:

  1. tell the token validation instance to listen only on localhost
    - this prevents anyone calling other endpoints of this instance
  2. return "404 Not Found" if request is not from localhost
    - this ensures that ONLY the main app instance can get _any_
      information about possible tokens' validity and hinders a
      brute-force attack, as we will likely rate-limit our main app's
      public endpoints.

Email Index
===========
In order to enable retrieving a user's information in order to validate
login credentials (email, password pair), we need to add a secondary
index to the User table. Here, for simplicity's sake, we used a DynamoDB
[Global Secondary Index (GSI)](http://amzn.to/2s4fz0n). In the future,
we may want to switch this to a Local Secondary Index once we add a
range key to the index (see [here](http://amzn.to/2s9bSSt) for a more
in-depth comparison).

Hidden Attributes
=================
We don't **ever** want a password or password hash to leave the
boundaries of our API. To ensure this, we introduce the notion of
"hidden attributes". These are represented in a `set` in `config.py`
called `HIDDEN_ATTRIBUTES`. In `handlers.py`, we add a function
`clean_response` which takes as input an entity dict and returns that
dict with any instances of a member of `HIDDEN_ATTRIBUTES` removed. As
long as we consistently update `HIDDEN_ATTRIBUTES` when such attributes
are introduced and faithfully apply this function to all returned entity
dicts, we won't accidentally leak any secrets.

Bug Fixes
=========
- Fix off-by-one error on offset checking, need to add test case.
- Change to `run_api.sh` that allows interative usage if `-d` isn't
  specified. Previously, it would suppress flask's debugger output and
  render pdb utterly useless. Now, we can drop into pdb at any point by
  adding the following where desired: `importd pdb; pdb.set_trace()`.

Miscellany
==========
- For debugging (and possibly testing) purposes, we allow the token
  `master_key` as a valid token for user with ID `1234` (which is utter
  nonsense). Right now, this is hardcoded, but we should consider moving
  it over to `config.py` so the tests can accesss it dynamically.
- because the test code has not yet been updated to use token-based
  authentication, ALL tests which don't expect a 401 will fail.
- changes `run_api.sh` to account for new app instance thread
- changes to other support scripts to account for necessary auth env
  vars and JSON attrs
- reduce maximum allowd paginated `limit` value from 65536 to 500.
- update dependency versions

New TODOs
=========
- see TODO's about email verification, commented-out function stubs
- see TODO about paginating scan result iterator in scalable/efficient
  fashion
- see TODO about `auth.py` refactor, that one should be pretty quick