One org per user, remove page limit default.
This commit modifies the User model to have a 1:1 mapping of user onto
organizations. The 1:1 nature of this should be bidirectional, allowing
each organization to have exactly one user (for now). Allowing a User to
belong to multiple organizations will not likely be useful in the
future, but perhaps having multiple users per org will be.

We also remove the default page limit of 10. This page limit is now
optional on the "Paginated" trait, and is valid on the interval [0,500].
We probably need to modify the tests around this.

To facilitate development and enhance logging, we now include in the
Exception's message the entity ID when updating an entity.