Add Task model.
At the moment, Tasks are created internally and only upon user account
creation. The list of tasks associated with an Organization (and thus,
due to the 1:1 mapping of User --> Organization, a unique User) are
initialized from a static "template" in the form of a python dict, which
can be found in `default_tasks.py`. The current tasks there are dummy
tasks, meant to be temporary placeholders until we figure out the
individual task names and associated workspaces.

We create a DDB GSI on org_id for the Task table.

We should consider creating an Organization on User account creation.