Fix all tests, add auth stubs, add BulbException.
In order to get Flask's `test_client` to automatically send our dummy
auth headers during testing, we created the inline
`AuthenticatedTestClientWrapper` class in the `client` fixture. We then
wrap the Flask app's `wsgi_app` attribute with this class's `__call__`
method to intercpet all outgoing API requests by the the app's test
client and inject our dummy auth token header (`master_key`). For this
to work, we had to capitalize all letters of the header name and prepend
a prefix of `HTTP_`. Weird.

Currently, we're passing in `port` as a keyword into `start_auth_thread`
(which, frankly, shouldn't even be used outside of `__init__`). This is
terrible, and I feel bad. And sorry. We need to move this and other shit
over to a test config file ASAP.

Stubbed out some auth test cases pretty well, now just have to implement
them...

We changed all attributes of the verbose name `organization` to
`org_id`. Further in this interest of standardization, we should at some
point change Organization's `users` to `user_ids`.

Bug Fixes
=========
- all the failing tests
- path variable `id` --> `task_id` for Task.{get,update,delete}
- typo from last commit, s/res_id/task_id where appropriate
- strangely, it seems that DDB serializes a value of "0.0" to "0",
  causing a mis-parsing of the latter as an int when the input (the
  former) was a decimal.
- calling `create_user` directly on a dict seems to modify that dict.
  This is usually fine, as it is connexion calling our handlers, and
  thus it doesn't matter if we have mutating side effects on the inputs
  to the handler. However, in testing, we re-use the objects passed into
  these functions as input, and thus are susceptible to breakage if our
  handlers have side effects. So, when we call `create_user` (which we
  _must_ do b/c of the password/HIDDEN_ATTRS special magic), we need to
  pass in _copies_ of our test fixtures.