Add gid to JWT auth token, enforce ownership.
This commit adds the `gid` claim to our auth JWT's to represent the
user's associatiated `org_id` (Organization). Again, this is slightly
superfluous due to the 1:1 relation of User <--> Organization, but will
allow us to associate multiple Users with an Organization later on.

We add a function `check_ownership` to `handlers.py` to verify that a
user is authorized to access the entity which they are specifically
trying to (i.e., the entity's `id` is specified as a URL path variable).
We do something similar for all `list_*` methods, but need to implement
it in the `list_entity` itself because no entity `id` is specified in
such calls.

Eventually, we should move this over to a standalone OAuth server and
use OAuth's scopes to determine authorization for particular resources.
Maybe?

Miscellania
===========
- upgrade connexion, Flask libraries and their dependencies
- add an `org_index` to every entity that keeps track of `org_id`
- we still have that awful `master_key` token that is all powerful and
  ever-abiding. It can see all and ruin everything. Namely, it's a
  massive security threat. But hey, this is 2017. That's normative.
- this scanning shit is still unfuckable inscalable. Wowzers.
- fix up and flesh-out some `TestReadEntity` tests around pagination
  query params; what happens when `offset == len(owned_entities)`? Well,
  counterintutively, it's `[]`.

  > but it should be a 400 because you're zero-indexed, thus
  > len(owned_entties) should be an invalid offset; one beyond the
  > bounds of the "array"!

  sure, but we're not fascists here. If we were to implement that, then
  specifying an offset of 0 for a would-be-empty list should throw a
  400. That's bullshit, and we avoid special cases in these parts, so
  the choice stands.
- throw 400 when non-`int` pagination param is passed.
- Let's throw around some interesting values in `default_tasks.py` so we
  can have something neat to show in the UI. Later, though, obvi
  `priority` will default to `0.0`, `status` --> `NOT_STARTED`, etc.
- move `master_key` `uid`/`gid` shit to `config.py`
- TODO: really need to implement this fucking authorization tests!!